LayoutTests/fast/forms/ios/validation-bubble-dismiss-on-tap.html

<!DOCTYPE html> <!-- webkit-test-runner [ useFlexibleViewport=true ] -->

<html>
<head>
    <meta name="viewport" content="width=device-width">
    <script src="../../../resources/js-test.js"></script>
    <script src="../form-validation.js"></script>
    <script>
    jsTestIsAsync = true;

    function getTapAtPointUIScript(x, y)
    {
        return `
            (function() {
                uiController.singleTapAtPoint(${x}, ${y}, function() {
                    uiController.uiScriptComplete('');
                });
            })();`
    }

    function doTest()
    {
        if (!testRunner.runUIScript)
            return;

        description("Tests that tapping the view dismisses the HTML form validation popover.");

        document.getElementById("testSubmitButton").click();

        getValidationMessage().then((_message) => {
            if (_message !==  "Select this checkbox")
                testFailed("Unexpected validation message: " + _message);

            testRunner.runUIScript(getTapAtPointUIScript(0, 0), function() {
                getValidationMessage().then((_message) => {
                    message = _message;
                    shouldBeEqualToString("message", ""); // Not visible.
                    finishJSTest();
                });
            });
        });
    }

    window.addEventListener('load', doTest, false);
    </script>
</head>
<body>
<form>
<input type="checkbox" id="testInput" required><input type="submit" id="testSubmitButton">
</form>
</body>
</html>
